pyproject.toml
[project]
name = "rec-sandbox"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shung-ich", email = "shung1taka@gmail.com" }
]
dependencies = [
    "torch-directml>=0.2.5.dev240914",
    "pandas>=2.2.3",
    "ipykernel>=6.29.5",
    "tqdm>=4.67.1",
    "scikit-learn>=1.6.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rec_sandbox"]
